+++ redis_consumer/consumers/multiplexed_consumer_test.py
